ENG-3805 Split scan transaction on migration delay
Post-Phoenix, migration requires a synchronous request
to polaris, which may fail for any number of reasons.

Crucially, if the parent under which the migrated object
is being moved was created in the same db transaction,
which is possible because ScanSession optimistically
groups multiple MightCreate into the same transaction,
the creation will be rolled-back and the error will
persist as the newly created folder will never be
advertised to polaris.

Change-Id: Ifa8d8d3f14ad8866a47f181b83df7de47df6b7e3