zephyr: fix ChannelDirectory/ChannelMonitor race
Zephyr transport expects at most one channel per DID, otherwise the
signalling service may dispatch the handhake message to the wrong
channel.

The extra synchronization done in ZephyrConnectionService was not
honored by ChannelMonitor, which made it possible for a race to
create two channels to the same device.

This would manifest as a rare case of handshake timeout causing a
~40s delay in establishing a connection, as occasionally seen in
CI.

Solved by moving synchronization from ZephyrConnectionService to
the ChannelDirectory instance.

Change-Id: Ia520d101c1c04d91e1fdfe89eb6c318dd5bfaad2