[ENG-2605] auto-upload appliance logs
This change allows admins automatically
upload appliance logs to dryad with a
subject and message. Logs are sent from
the support address.

Compression, a test for dryad
availability, and the email are handled
handled in the foreground - so the user
knows if they fail and can reach out
manually. Upload is jobbed off to a
celery worker - it can fail without
notifying the user, because we will
have the email, at least.

Change-Id: I2b4af81bab4f5ce56f1f8c4aefdbf9749b0c14ce