SUPPORT-2208 Fix ncwsv check
When a content request is received and the requester already
has the version the can be served by the requestee, treat it
as ncwsv.

Failure to do so could lead to repeated transfer attempts,
much noise in the logs and would delay bloom filter cleanup
which would cause more false positives and further slow
sync.

Change-Id: Ia4434402e18684e9416c9e802eca1e375f661bf1