[ENG-4190] When loading, hide 'No files found'
In shelob-backed views the 'No files found'
message is replaced with 'Loading file list'
when the request for the object list hasn't
yet completed or finally failed.

Furthermore, file-list.html is changed so the
loading message persists while shelob retries
the requests necessary to render pages for
links to individual files.

Change-Id: I9662166b495ecf12fb79ec980ece6a4196045ebe