waldo: reserve type byte for each row
At steady state, most objects within a store will be present
on identical sets of devices and the version of a given object
will eventually be the same on all devices on which it is
present.

This leaves a lot of room for significantly more space-efficient
row encodings. For instance one could separate DIDs and versions
into two blocks instead of having them be interleaved. A sorted
DID set could then be efficiently compressed via a dictionary and
a version list would be amenable to run-length or entropy encoding.

Change-Id: I06cd5475f558c3945c1fb0380563544bb58180dc