API upload: send reply after polaris sub
First step to make sure successful uploads are immediately
available for download.

Next step will be to have polaris advertise location on
behalf of client on successful new version submission.

Note that both of these are best-effort. We do not fail
the upload if the submission fails or times out.

Change-Id: Ic5c7156cc2c1c73260137506fca8adb46f028b93