API upload: send reply after waldo advertisement
It would be possible to avoid a client<->appliance roundtrip there
but it would introduce significant complexity on the server side
as polaris would need to make the location advertisement to waldo
on behalf of the device.

Take the simple route for now and revisit later if the increased
latency turns out to be a problem.

Change-Id: I56d74c165d15cc4306d2fedd3a4f0f05b07a6bd4