ENG-3944 MSI installers can auto-update
NB: This will NOT update the MSI itself as "Program Files"
is unlikely to be writable this is not even attempted.

Instead if will use the regular update process, which consists
in installing the regular exe package in the per-user approot
and chainloading to the installer.

Testing on Win7 shows that no duplicate start menu entry are
visible. Instead the most recent one (exe installer) take
precedence over the previous one (msi installer) which is fine.

The main drawback of this approach is uninstallation as the
NSIS uninstaller will NOT uninstall the MSI so files will
remain in "Program Files" and the previously hidden start
menu entry will resurface. This is a little annoying but it
does make it easy for the user to uninstall the MSI as well.

Change-Id: Ib5ec099f56905d02a194a562b14af755cac1de69