[ENG-4322] AeroFS users can invite to lizard
After this change, AeroFS users on the
internal app will be able to register users
for AeroFS accounts.

Users will still need to follow a link to
activate their accounts and choose a password.

Marketing asked that this hit the existing
Pardot endpoint for sales, so all data
required by that endpoint are collected in
this workflow.

Change-Id: Iaae20e6ba85bcf49ad36786ac1fb20b3ebbab446