ENG-4452 allow cert auth to obtain OAuth token
The resulting token has the DID of the cert used to obtain it,
instead of a random MDID. This allows audit/activity log to
correctly reflect the origin of meta changes.

NB the page listing mobile devices might need to be updated to
handle this new use case more gracefully.

Change-Id: I77ad5e42828f50627502e6603ba51afca9ed4b12