Fix cert revocation
Certs were never revoked because the default value for a timestamp
is null, not zero. This was not caught in tests either because of
gradle not failing properly on parallel runs or because mysql and
mariadb behavior is subtly different in this case.

Check for unlinked device instead of revoked serials as unlink is
the only way for certs to be revoked.

Change-Id: Ia13c2aecd81c3798c20b7e3f067678f70b647bd7