ENG-4739 Wait for ACL sync on file upload
When a file is uploaded immediately after a new shared folder is
created, there is a race between the API request and ACL
propagation which could result in the first few upload attempts
failing.

We're already waiting for the OID to appear so it makes sense to
use a similar approach for the store.

Change-Id: I7d2d834be6b3ac92e4175af86be967f31d72dbf2