SUPPORT-2230: relax cleanup invariant
TS makes everything more complex than it needs to be,
can't wait for it to be deprecated.

Relax an invariant to prevent a crash loop in some
corner cases involving pre-Phoenix shared folders
on a TS. It *may* result in some files being moved
to sync history instead of migrated if the TS was
ahead of other peers but that behavior is not
actually a regression from pre-Phoenix behavior
and sync history should allow recovery.

Change-Id: Ibb25a665ea94d79b63e16c5c787845a96b5871f0