identity: no SAML? no tomcat!
Avoid spawning a tomcat instance if the authenticator
isn't either OpenID or SAML.

Instead a netcat process is started to keep the probe
happy (said probe is adapted to test for an open port
instead of making an HTTP request).

This shaves upwards of 60Mb off the appliance memory
footprint, which is especially significant in an HPC
context.

Change-Id: I1da0fccc69a2863e7d7c8863547a42d20ce8e6d6