ENG-4815: Native loading bar for OS X
This change introduces a native process for OS X
that displays a progress bar while the golang updater
downloads content to approot.

Upater passes the file count to the native process it
spawns. Updater also passes its progress to the monitor
via stdin.

Updater kills the native UI if it errors or detects that
the download is complete.

Progress is computed based on the number of the files in
the manifest that have been processed.

Updaters and osx binary to follow in a later patchset.

Change-Id: I0925f63543a46bf1dacbc29510439da3c28c7366