Undo fetch images on the fly
A big issue with fetching images on the fly is if we are not able to docker pull
for some reason, or if docker is busted for some reason on the appliance then
we are in bit of a pickle because we never get to getty screen. Undoing it will
atleast make sure we get to the getty screen and have root access and such.
Revisit this when and if we are proactive about shipping lightweight images

Change-Id: I0f8dd932e0a8cda64a0fd9999bacc0ad097241df