Fix handling of meta changes in migrated src tree
Local changes can happen before reception of the
DELETE_CHILD transform for the migrated subtree.

These changes will never be accepted by polaris as
the migrated source tree is permanently locked to
avoid races so it is up to the daemon to fix or
discard them appropriately.

The old code would fail to cleanup some such changes
which could lead to crash loops or similar weirdness
as in the case of monumentfs where a temporary file
was created and immediately deleted under the migrated
tree but the second local meta change was not cleaned
up, leading to a crash loop in MetaChangeSubmitter.

Change-Id: I48ab89824d085cb83a4357256e964866e1a30cf9