ENG-6357: Fix redis cache before restart
Attempt to fix the redis cache before restarting on error.

Tested by appending garbage data to /data/appendonly.aof, then restarting
the redis container. The logs show the first startup fails, then the script
repairs the cache, and the second startup succeeds. Repeating this a second
time shows the old backup file is untouched, and a new backup file is
created, as expected. The backup files have a timestamp in their file name
to avoid overwriting, and are never automatically deleted.

Change-Id: If2c70df48afe21c43085f0eaaeefeba825e540df