SUPPORT-2249 daemon: fix overzealous assert
When moving an object under the root, polaris response
will refer to the root folder by the SID, which needs
to be mapped back to OID.ROOT for daemon use.

This is annoying to test in unit tests and was not hit
by system tests because all syncdet cases use per-run
folders to avoid unintended name conflicts.

Change-Id: Id41ecb1b93249df91d8aea0b2c3a22681beacc57