SUPPORT-2254 daemon: misc block storage fixes
Simplify block refcount by removing use of the state field
which was rendered essentially useless by the introduction
of background block upload. This should hopefully solve the
AE seen by Alibaba which triggered a cascading failure in
error handling.

Add guards in close() methods to avoid invariant violations
in finally blocks and the resulting cascading failure.

Add more detailed error logging to help identifying the root
cause of the triggering failure in case it is not actually
fixed.

Change-Id: Ibaef9b03dcb6716900bc7e4eba2fd68cce1ff8d1