daemon: Fix false conflict and resulting no-sync
As seen on Yuri's MBP (did 97356118EAE74913AC130BD2093B643D)

Logs do not go back far enough to accurately reconstruct what
happened but the current state of the db suggests the following
scenario:

A file is modified
Daemon submit content change to polaris
The same file is modified *after* the request is sent to polaris
Either the response is lost due to network issues or the transform
fetch wins the race with the response. At this point the daemon
records the existence of a new polaris-blessed content version but
it *cannot* fast-forward to it because its local content no longer
matches the hash.

This leaves the daemon unable to service content requests for this
file from other peers, and therefore unable to get that transient
version from any peer. The daemon is also unable to submit its most
recent version resulting in a no-sync. Modifying that file on another
machine would also result in a no-sync as it wouldn't be able to either
submit the new change or create a conflict branch.

Unfortunately this bad state is largely invisible to the user as it
doesn't get to the point of creating a conflict branch. This suggests
the need for an extended form of unsyncable/conflicting files that
would take repeated protocol errors into account.

Upgrade will *not* magically fix affected files. The only way to do
that is to either delete/re-create the affected files or go for a
full unlink/reinstall of *all* clients with a least one affected file.

Change-Id: I5f5cc03d4baa369aef4da59d19c10b7cd9e63f42