SUPPORT-2261 TS: Fix OOM in BlockStorage
As seen by alibaba, where it caused a crash loop.

Slow backend and lots of sync activity would lead to the accumulation
of a large number of files in the upload directory (524k blocks in the
alibaba case).

Listing the entire contents of the directory would in turn take up a
ridiculous 50Mb of RAM, putting the TS within spitting distance of the
64Mb heap limit.

Switch to java 7 streaming API to incrementally list the contents of the
upload dir, thereby keeping memory usage low.

Change-Id: I93d6872b0855d3410ed66b1c77b6d7ddb49daad2