SUPPORT-2260 updater: bump http fetch timeout to 5min
This is a band-aid. Ideally we wouldn't have a hard timeout on
fetching an entire thing but a timeout on reading more data.

On a related note, it would be nice (but non-trivial) to be
able to reuse downloaded files across failed updater runs,
the pinnacle being to resume incomplete download but that
would be even harder so probably not worth the trouble to
be honest.

Another thing worth considering would be to re-introduce
fat installers now that repackaging is gone. They would
allow a nicer UX for the initial install and we'd still
have the benefit of fine-grained upgrades.

Change-Id: I2f18fb9c1792856a8a79c8a59673a0002ce22670