More robust Phoenix conversion
"Wait, what? Are we still doing this?" you ask

Why, yes, we very much are. Or at least, our dearest customers are.

Credit Karma recently migrated from the positively ancient 1.0.19
That was not merely pre-Phoenix but (gasp!) pre-docker as well!

At first glance, things went pretty well, considering that the
two versions are more than 2 years apart.The Team Server did get
stuck on a repeated failure to submit conversion ops as one of
the folders it tried to submit was rejected by polaris due to
a name conflict. It is quite surprising that this case was not
handled, not triggered in testing and (to the best of our
knowledge) did not occur in the wild (it's quite possible it
*did* occur and was simply handled by an unlink/reinstall).

This changes addresses two related but distinct issues:
 - repeated failures in a store should not affect conversion
   progress in other stores
 - name conflicts should be handled in the client to allow
   conversion to proceed. Conflict resolution is deferred
   to the regular Phoenix fetch/submit machinery and the
   conflicting object simply omitted from the conversion
   ops going forward.

Change-Id: I092cd4c6c5ba6cc6071d6ee0a848d1baed58aa4b