Enable APFS on macOS
This is the default starting with 10.13 (High Sierra) and SSD volumes
are automatically re-formatted upon upgrade from an older version of
the OS.

While HFS+ was only using 32bit inode numbers, the native code had
been recompiled a few years ago to use 64bit ino_t so that does not
cause any problems.

However, the upgrade will invalidate all inode numbers stored in the
DB and cause a lot of extra work to update things upon upgrade, and
potentially lose track of a few moves that were not properly handled
before the upgrade. This is roughly the same thing that would happen
if, for instance, the user installed a new drive into a machine and
migrated all the data over with the machine turned off, except this
scenario goes from "very unlikely" to "affect all users".

Change-Id: I730502c66dd421495e5fd5dfd7a795a007744f96