ScanSession: reduce memory usage
Use a set of Path instead of PathCombo to keep track of visited
folders. This ensures the PathCombo objects (and through them
the full absolute paths) can be garbage collected sooner.

Ideally we wouldn't keep a full set of Path just to avoid double
visit as that still risks causing OOMs. However, Path is better
a de-duping the underlying strings so this should already be a
significant improvement.

Change-Id: I4484e1d7ac55f195a05675677e5069026a60d7d1