Mitigate rare daemon zombification
As seen at spainwilliams

OOM would sometimes results in some threads being killed but
the process remaining up. At which point the progress watcher
would kick in and find no progress being done. It would then
try to submit a defect report but due to the previous OOM this
would fail with a NoClassDefFoundError when trying to create
the auto-defect. This would in turn inhibit the fatal() call
and lead to a repeated flooding of the daemon logs with stack
traces.

Change-Id: I33be236585fc73f4afd235b37bbd4ebb209f511d