ENG-8264 ApplyChange: relax version jump invariant
Allow version jump when the transform comes with a migrant OID

This is to account for what looks to be some weird old behavior
in the polaris migration logic, as seen in the transform log of
Fabrique's appliance. It is hoped that the root cause was fixed
by a series of changes by Rudy in May 2016, which is conveniently
shortly after the timestamps of the offending transforms in the
mysql dump obtained from Fabrique.

Change-Id: Id9674dc03386b98131e86cb0b7b274da9548a02c