docker/dev: use colima instead of vbox/docker-machine
docker-machine was EOLed years ago
colima looks to be the most promising replacement

this move requires change to the package caching infrastructure
which was based on a really old DNS override setup that pre-dated
the introduction of docker's build-in DNS resolver on user-created
bridge networks.

drop apt-cacher-ng / alpinx, which relied on rawdns for transparent
proxying. we can revisit those later if build times are a concern.

bump devpi to a more recent version and use build-time network
attachment to resolve devpi

As a side bonus, using devpi proxy for pip installs allows us to
keep supporting obsolete versions of pip on py2.7 that otherwise
fail to fetch packages due to change in PyPI's url and/or outdated
certificates